requirements.txt
# requirements.txt
discord.py
python-dotenv
SQLAlchemy
psycopg2-binary
aiohttp